requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.7
django-debug-toolbar==3.2
django-filter==2.4.0
gunicorn==20.1.0
mysqlclient==2.0.3
Pillow==8.2.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
